+++ src/primitives/option.cpp
@@ -65,23 +65,35 @@ std::shared_ptr<Obj>
 Option::binop(Token *op, std::shared_ptr<Obj> &other) {
     ASSERT_BINOP_COMPAT(this, other.get(), op);
 
-    if (op->type() != TokenType::Double_Equals) {
-        std::string msg = "the only support binary operations on option types is equality `==`";
+    auto other2 = dynamic_cast<Option *>(other.get());
+    if (op->type() == TokenType::Double_Equals) {
+        if (this->is_none() && other2->is_none())
+            return std::make_shared<Bool>(true);
+
+        if (this->is_some() && other2->is_none())
+            return std::make_shared<Bool>(false);
+
+        if (this->is_none() && other2->is_some())
+            return std::make_shared<Bool>(false);
+
+        return std::make_shared<Bool>(this->value()->eq(other2->value()));
+    }
+    else if (op->type() == TokenType::Bang_Equals) {
+        if (this->is_none() && other2->is_none())
+            return std::make_shared<Bool>(false);
+
+        if (this->is_some() && other2->is_none())
+            return std::make_shared<Bool>(true);
+
+        if (this->is_none() && other2->is_some())
+            return std::make_shared<Bool>(true);
+
+        return std::make_shared<Bool>(!this->value()->eq(other2->value()));
+    }
+    else {
+        std::string msg = "invalid operator for binary operation `"+op->lexeme()+"` on option type";
         throw InterpreterException(msg);
     }
-
-    auto other2 = dynamic_cast<Option *>(other.get());
-
-    if (this->is_none() && other2->is_none())
-        return std::make_shared<Bool>(true);
-
-    if (this->is_some() && other2->is_none())
-        return std::make_shared<Bool>(false);
-
-    if (this->is_none() && other2->is_some())
-        return std::make_shared<Bool>(false);
-
-    return this->value()->binop(op, other2->value());
 }
 
 bool